Create `underscore` module
Module `underscore` contains common pandas operations and utilities and
creates pandas api extensions `_` as the accessor to those utilities.

- move `str_join` method to that module


new file:   thoth/lab/underscore.py
modified:   thoth/lab/utils.py
